pyproject.toml
[project]
name = "autotrust"
version = "0.1.0"
description = "Masubi: Autonomous email trust scoring research loop"
requires-python = ">=3.12"
dependencies = [
    "anthropic",
    "openai",
    "ollama",
    "python-dotenv",
    "pydantic",
    "pyyaml",
    "gitpython",
    "httpx",
    "rich",
    "structlog",
    "datasets",
    "scikit-learn",
    "torch>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]
dashboard = [
    "gradio>=5.0",
    "plotly>=5.0",
    "pandas>=2.0",
]
export = [
    "llama-cpp-python>=0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["autotrust"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint.per-file-ignores]
# load_dotenv() must run before importing modules that read env vars
"run_loop.py" = ["E402"]
